server.js
Save this file in the directory you want to use as your server. I used desktop. 

Run from command prompt using " node server.js" from the directory this file is in. (I right click desktop and open the shell from there)

This server script requires express, body-parser, cors, and helmet to run. In command prompt type "npm install helmet cors express body-parser" in the directory this server will be run.

once node server.js is typed, the command prompt will show "listening on port 3000" or whatever you choose in the code.

for troubleshooting, when the fullscan is run on BB, the info will print on BB and the command prompt, before displaying on the local webpage.